@@ -1,24 +1,22 @@
-import { TestBed, inject } from '@angular/core/testing';
-import { PlatformRef, NgModule, CompilerFactory, NgZone } from '@angular/core';
-import { FirebaseApp, AngularFireModule } from './public_api';
-import { Subscription, Observable, Subject, of } from 'rxjs';
+import { TestBed } from '@angular/core/testing';
+import { CompilerFactory, NgModule, NgZone, PlatformRef } from '@angular/core';
+import { AngularFireModule, FirebaseApp } from './public_api';
+import { Observable, of, Subject } from 'rxjs';
 import { COMMON_CONFIG } from '../test-config';
 import { BrowserModule } from '@angular/platform-browser';
-import { database } from 'firebase/app';
-import { ɵZoneScheduler, ɵkeepUnstableUntilFirstFactory, ɵAngularFireSchedulers } from './angularfire2';
-import { ɵPLATFORM_BROWSER_ID, ɵPLATFORM_SERVER_ID } from '@angular/common';
-import { tap  } from 'rxjs/operators';
+import firebase from 'firebase/app';
+import { ɵAngularFireSchedulers, ɵkeepUnstableUntilFirstFactory, ɵZoneScheduler } from './angularfire2';
+import { tap } from 'rxjs/operators';
 import { TestScheduler } from 'rxjs/testing';
 import { rando } from '../firestore/utils.spec';
 
 describe('angularfire', () => {
-  let subscription:Subscription;
   let app: FirebaseApp;
-  let rootRef: database.Reference;
-  let questionsRef: database.Reference;
-  let listOfQuestionsRef: database.Reference;
+  let rootRef: firebase.database.Reference;
+  let questionsRef: firebase.database.Reference;
+  let listOfQuestionsRef: firebase.database.Reference;
   let defaultPlatform: PlatformRef;
-  let appName:string;
+  let appName: string;
 
   beforeEach(() => {
 
@@ -28,27 +26,21 @@ describe('angularfire', () => {
       imports: [AngularFireModule.initializeApp(COMMON_CONFIG, appName)]
     });
 
-    inject([FirebaseApp, PlatformRef], (_app: FirebaseApp, _platform: PlatformRef) => {
-      app = _app;
-      rootRef = app.database!().ref();
-      questionsRef = rootRef.child('questions');
-      listOfQuestionsRef = rootRef.child('list-of-questions');
-      defaultPlatform = _platform;
-    })();
-
+    app = TestBed.inject(FirebaseApp);
+    defaultPlatform = TestBed.inject(PlatformRef);
+    rootRef = app.database().ref();
+    questionsRef = rootRef.child('questions');
+    listOfQuestionsRef = rootRef.child('list-of-questions');
   });
 
   afterEach(() => {
-    rootRef.remove()
-    if(subscription && !subscription.closed) {
-      subscription.unsubscribe();
-    }
+    rootRef.remove();
     app.delete();
   });
 
   describe('ZoneScheduler', () => {
     it('should execute the scheduled work inside the specified zone', done => {
-      let ngZone = Zone.current.fork({
+      const ngZone = Zone.current.fork({
         name: 'ngZone'
       });
       const rootZone = Zone.current;
@@ -64,11 +56,11 @@ describe('angularfire', () => {
     });
 
     it('should execute nested scheduled work inside the specified zone', done => {
-      const testScheduler = new TestScheduler(null!);
+      const testScheduler = new TestScheduler(null);
       testScheduler.run(helpers => {
         const outsideAngularScheduler = new ɵZoneScheduler(Zone.current, testScheduler);
 
-        let ngZone = Zone.current.fork({
+        const ngZone = Zone.current.fork({
           name: 'ngZone'
         });
 
@@ -97,13 +89,13 @@ describe('angularfire', () => {
               helpers.flush();
               done();
               expect(callbacksRan).toEqual(3);
-            })
+            });
           });
           helpers.flush();
         });
       });
-    })
-  })
+    });
+  });
 
   describe('keepUnstableUntilFirstFactory', () => {
     let schedulers: ɵAngularFireSchedulers;
@@ -121,10 +113,10 @@ describe('angularfire', () => {
         runTask: insideZone.run.bind(insideZone)
       } as NgZone;
       schedulers = new ɵAngularFireSchedulers(ngZone);
-    })
+    });
 
     it('should re-schedule emissions asynchronously', done => {
-      const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(schedulers, ɵPLATFORM_SERVER_ID);
+      const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(schedulers);
 
       let ran = false;
       of(null).pipe(
@@ -133,38 +125,39 @@ describe('angularfire', () => {
       ).subscribe(() => {
         expect(ran).toEqual(true);
         done();
-      }, () => fail("Should not error"));
+      }, () => fail('Should not error'));
 
       expect(ran).toEqual(false);
     });
 
-    [ɵPLATFORM_SERVER_ID, ɵPLATFORM_BROWSER_ID].map(platformId =>
-      it(`should subscribe outside angular and observe inside angular (${platformId})`, done => {
-        const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(schedulers, platformId);
+    it(`should subscribe outside angular and observe inside angular`, done => {
 
-        insideZone.run(() => {
-          new Observable(s => {
-            expect(Zone.current).toEqual(outsideZone);
-            s.next("test");
-          }).pipe(
-            keepUnstableOp,
-            tap(() => {
-              expect(Zone.current).toEqual(insideZone);
-            })
-          ).subscribe(() => {
+      const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(schedulers);
+
+      insideZone.run(() => {
+        new Observable(s => {
+          expect(Zone.current).toEqual(outsideZone);
+          s.next('test');
+        }).pipe(
+          keepUnstableOp,
+          tap(() => {
             expect(Zone.current).toEqual(insideZone);
-            done();
-          }, err => {
-            fail(err);
-          });
+          })
+        ).subscribe(() => {
+          expect(Zone.current).toEqual(insideZone);
+          done();
+        }, err => {
+          fail(err);
         });
-      })
-    );
+      });
 
-    it('should block until first emission on server platform', done => {
-      const testScheduler = new TestScheduler(null!);
+    });
+
+    it('should block until first emission', done => {
+      const testScheduler = new TestScheduler(null);
       testScheduler.run(helpers => {
         const outsideZone = Zone.current;
+        // tslint:disable-next-line:no-string-literal
         const taskTrack = new Zone['TaskTrackingZoneSpec']();
         const insideZone = Zone.current.fork(taskTrack);
         const trackingSchedulers: ɵAngularFireSchedulers = {
@@ -175,14 +168,18 @@ describe('angularfire', () => {
             runTask: insideZone.run.bind(insideZone)
           } as NgZone,
           outsideAngular: new ɵZoneScheduler(outsideZone, testScheduler),
-          insideAngular: new ɵZoneScheduler(insideZone, testScheduler),
+          insideAngular: new ɵZoneScheduler(insideZone, testScheduler)
         };
-        const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(trackingSchedulers, ɵPLATFORM_SERVER_ID);
+        const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(trackingSchedulers);
 
         const s = new Subject();
         s.pipe(
-          keepUnstableOp,
-        ).subscribe(() => { }, err => { fail(err); }, () => { });
+          keepUnstableOp
+        ).subscribe(() => {
+        }, err => {
+          fail(err);
+        }, () => {
+        });
 
         // Flush to ensure all async scheduled functions are run
         helpers.flush();
@@ -194,47 +191,17 @@ describe('angularfire', () => {
         s.next(123);
         helpers.flush();
 
-        // Should not be blocked after first item
-        expect(taskTrack.macroTasks.length).toBe(0);
+        // TODO drop this, it's to work around my 15ms timeout hack
+        setTimeout(() => {
+          // Should not be blocked after first item
+          expect(taskTrack.macroTasks.length).toBe(0);
+          done();
+        }, 150);
 
-        done();
       });
-    })
+    });
 
-    it('should not block on client platform', done => {
-      const testScheduler = new TestScheduler(null!);
-      testScheduler.run(helpers => {
-        const outsideZone = Zone.current;
-        const taskTrack = new Zone['TaskTrackingZoneSpec']();
-        const insideZone = Zone.current.fork(taskTrack);
-        const trackingSchedulers: ɵAngularFireSchedulers = {
-          ngZone: {
-            run: insideZone.run.bind(insideZone),
-            runGuarded: insideZone.runGuarded.bind(insideZone),
-            runOutsideAngular: outsideZone.runGuarded.bind(outsideZone),
-            runTask: insideZone.run.bind(insideZone)
-          } as NgZone,
-          outsideAngular: new ɵZoneScheduler(outsideZone, testScheduler),
-          insideAngular: new ɵZoneScheduler(insideZone, testScheduler),
-        };
-        const keepUnstableOp = ɵkeepUnstableUntilFirstFactory(trackingSchedulers, ɵPLATFORM_BROWSER_ID);
-
-        const s = new Subject();
-        s.pipe(
-          keepUnstableOp,
-        ).subscribe(() => { }, err => { fail(err); }, () => { });
-
-        // Flush to ensure all async scheduled functions are run
-        helpers.flush();
-
-        // Zone should not be blocked
-        expect(taskTrack.macroTasks.length).toBe(0);
-        expect(taskTrack.microTasks.length).toBe(0);
-
-        done();
-      });
-    })
-  })
+  });
 
   describe('FirebaseApp', () => {
 
@@ -253,13 +220,15 @@ describe('angularfire', () => {
           imports: [
             AngularFireModule.initializeApp(COMMON_CONFIG, appName),
             BrowserModule
-          ]})
+          ]
+        })
         class MyModule {
-          ngDoBootstrap() {}
+          ngDoBootstrap() {
+          }
         }
 
         const compilerFactory: CompilerFactory =
-            defaultPlatform.injector.get(CompilerFactory, null);
+          defaultPlatform.injector.get(CompilerFactory, null);
         const moduleFactory = compilerFactory.createCompiler().compileModuleSync(MyModule);
 
         defaultPlatform.bootstrapModuleFactory(moduleFactory)
@@ -267,10 +236,10 @@ describe('angularfire', () => {
             const ref = moduleRef.injector.get(FirebaseApp);
             expect(ref.name).toEqual(app.name);
           }).then(done, e => {
-            fail(e);
-            done()
-          });
-      })
+          fail(e);
+          done();
+        });
+      });
 
     }
   });
