@@ -1,21 +1,20 @@
-import { database } from 'firebase/app';
-import { FirebaseApp, AngularFireModule } from '@angular/fire';
+import firebase from 'firebase/app';
+import { AngularFireModule, FirebaseApp } from '@angular/fire';
 import { AngularFireDatabase, AngularFireDatabaseModule, listChanges, URL } from '../public_api';
-import { TestBed, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { COMMON_CONFIG } from '../../test-config';
-import { skip, take, tap } from 'rxjs/operators';
+import { skip, take } from 'rxjs/operators';
 import 'firebase/database';
 import { rando } from '../../firestore/utils.spec';
 
 describe('listChanges', () => {
   let app: FirebaseApp;
   let db: AngularFireDatabase;
-  let ref: (path: string) => database.Reference;
+  let ref: (path: string) => firebase.database.Reference;
   let batch = {};
-  const items = [{ name: 'zero' }, { name: 'one' }, { name: 'two' }].map((item, i) => ( { key: i.toString(), ...item } ));
-  Object.keys(items).forEach(function (key, i) {
-    const itemValue = items[key];
-    batch[i] = itemValue;
+  const items = [{ name: 'zero' }, { name: 'one' }, { name: 'two' }].map((item, i) => ({ key: i.toString(), ...item }));
+  Object.keys(items).forEach((key, i) => {
+    batch[i] = items[key];
   });
   // make batch immutable to preserve integrity
   batch = Object.freeze(batch);
@@ -30,11 +29,10 @@ describe('listChanges', () => {
         { provide: URL, useValue: 'http://localhost:9000' }
       ]
     });
-    inject([FirebaseApp, AngularFireDatabase], (app_: FirebaseApp, _db: AngularFireDatabase) => {
-      app = app_;
-      db = _db;
-      ref = (path: string) => db.database.ref(path);
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    db = TestBed.inject(AngularFireDatabase);
+    ref = (path: string) => db.database.ref(path);
   });
 
   afterEach(() => {
@@ -46,7 +44,7 @@ describe('listChanges', () => {
     it('should stream value at first', (done) => {
       const someRef = ref(rando());
       const obs = listChanges(someRef, ['child_added']);
-      const sub = obs.pipe(take(1)).subscribe(changes => {
+      obs.pipe(take(1)).subscribe(changes => {
         const data = changes.map(change => change.payload.val());
         expect(data).toEqual(items);
       }).add(done);
@@ -56,7 +54,7 @@ describe('listChanges', () => {
     it('should process a new child_added event', done => {
       const aref = ref(rando());
       const obs = listChanges(aref, ['child_added']);
-      const sub = obs.pipe(skip(1),take(1)).subscribe(changes => {
+      obs.pipe(skip(1), take(1)).subscribe(changes => {
         const data = changes.map(change => change.payload.val());
         expect(data[3]).toEqual({ name: 'anotha one' });
       }).add(done);
@@ -67,7 +65,7 @@ describe('listChanges', () => {
     it('should stream in order events', (done) => {
       const aref = ref(rando());
       const obs = listChanges(aref.orderByChild('name'), ['child_added']);
-      const sub = obs.pipe(take(1)).subscribe(changes => {
+      obs.pipe(take(1)).subscribe(changes => {
         const names = changes.map(change => change.payload.val().name);
         expect(names[0]).toEqual('one');
         expect(names[1]).toEqual('two');
@@ -79,7 +77,7 @@ describe('listChanges', () => {
     it('should stream in order events w/child_added', (done) => {
       const aref = ref(rando());
       const obs = listChanges(aref.orderByChild('name'), ['child_added']);
-      const sub = obs.pipe(skip(1),take(1)).subscribe(changes => {
+      obs.pipe(skip(1), take(1)).subscribe(changes => {
         const names = changes.map(change => change.payload.val().name);
         expect(names[0]).toEqual('anotha one');
         expect(names[1]).toEqual('one');
@@ -93,7 +91,7 @@ describe('listChanges', () => {
     it('should stream events filtering', (done) => {
       const aref = ref(rando());
       const obs = listChanges(aref.orderByChild('name').equalTo('zero'), ['child_added']);
-      obs.pipe(skip(1),take(1)).subscribe(changes => {
+      obs.pipe(skip(1), take(1)).subscribe(changes => {
         const names = changes.map(change => change.payload.val().name);
         expect(names[0]).toEqual('zero');
         expect(names[1]).toEqual('zero');
