@@ -1,17 +1,17 @@
 import { fromRef } from '../observable/fromRef';
-import { Observable, of, merge, SchedulerLike } from 'rxjs';
+import { merge, Observable, of, SchedulerLike } from 'rxjs';
 
-import { DatabaseQuery, ChildEvent, SnapshotAction } from '../interfaces';
+import { ChildEvent, DatabaseQuery, SnapshotAction } from '../interfaces';
 import { isNil } from '../utils';
 
-import { switchMap, distinctUntilChanged, scan } from 'rxjs/operators';
+import { distinctUntilChanged, scan, switchMap } from 'rxjs/operators';
 
-export function listChanges<T=any>(ref: DatabaseQuery, events: ChildEvent[], scheduler?: SchedulerLike): Observable<SnapshotAction<T>[]> {
+export function listChanges<T = any>(ref: DatabaseQuery, events: ChildEvent[], scheduler?: SchedulerLike): Observable<SnapshotAction<T>[]> {
   return fromRef(ref, 'value', 'once', scheduler).pipe(
     switchMap(snapshotAction => {
       const childEvent$ = [of(snapshotAction)];
       events.forEach(event => childEvent$.push(fromRef(ref, event, 'on', scheduler)));
-      return merge(...childEvent$).pipe(scan(buildView, []))
+      return merge(...childEvent$).pipe(scan(buildView, []));
     }),
     distinctUntilChanged()
   );
@@ -19,8 +19,8 @@ export function listChanges<T=any>(ref: DatabaseQuery, events: ChildEvent[], sch
 
 function positionFor<T>(changes: SnapshotAction<T>[], key) {
   const len = changes.length;
-  for(let i=0; i<len; i++) {
-    if(changes[i].payload.key === key) {
+  for (let i = 0; i < len; i++) {
+    if (changes[i].payload.key === key) {
       return i;
     }
   }
@@ -28,11 +28,11 @@ function positionFor<T>(changes: SnapshotAction<T>[], key) {
 }
 
 function positionAfter<T>(changes: SnapshotAction<T>[], prevKey?: string) {
-  if(isNil(prevKey)) {
+  if (isNil(prevKey)) {
     return 0;
   } else {
     const i = positionFor(changes, prevKey);
-    if( i === -1) {
+    if (i === -1) {
       return changes.length;
     } else {
       return i + 1;
@@ -41,7 +41,7 @@ function positionAfter<T>(changes: SnapshotAction<T>[], prevKey?: string) {
 }
 
 function buildView(current, action) {
-  const { payload, type, prevKey, key } = action;
+  const { payload, prevKey, key } = action;
   const currentKeyPosition = positionFor(current, key);
   const afterPreviousKeyPosition = positionAfter(current, prevKey);
   switch (action.type) {
@@ -49,7 +49,7 @@ function buildView(current, action) {
       if (action.payload && action.payload.exists()) {
         let prevKey = null;
         action.payload.forEach(payload => {
-          const action = {payload, type: 'value', prevKey, key: payload.key};
+          const action = { payload, type: 'value', prevKey, key: payload.key };
           prevKey = payload.key;
           current = [...current, action];
           return false;
@@ -60,14 +60,14 @@ function buildView(current, action) {
       if (currentKeyPosition > -1) {
         // check that the previouskey is what we expect, else reorder
         const previous = current[currentKeyPosition - 1];
-        if ((previous && previous.key || null) != prevKey) {
+        if ((previous && previous.key || null) !== prevKey) {
           current = current.filter(x => x.payload.key !== payload.key);
           current.splice(afterPreviousKeyPosition, 0, action);
         }
       } else if (prevKey == null) {
         return [action, ...current];
       } else {
-        current = current.slice()
+        current = current.slice();
         current.splice(afterPreviousKeyPosition, 0, action);
       }
       return current;
@@ -76,9 +76,9 @@ function buildView(current, action) {
     case 'child_changed':
       return current.map(x => x.payload.key === key ? action : x);
     case 'child_moved':
-      if(currentKeyPosition > -1) {
+      if (currentKeyPosition > -1) {
         const data = current.splice(currentKeyPosition, 1)[0];
-        current = current.slice()
+        current = current.slice();
         current.splice(afterPreviousKeyPosition, 0, data);
         return current;
       }
