@@ -1,22 +1,21 @@
-import { database } from 'firebase/app';
-import { FirebaseApp, AngularFireModule } from '@angular/fire';
-import { AngularFireDatabase, AngularFireDatabaseModule, snapshotChanges, ChildEvent, URL } from '../public_api';
-import { TestBed, inject } from '@angular/core/testing';
+import firebase from 'firebase/app';
+import { AngularFireModule, FirebaseApp } from '@angular/fire';
+import { AngularFireDatabase, AngularFireDatabaseModule, ChildEvent, snapshotChanges, URL } from '../public_api';
+import { TestBed } from '@angular/core/testing';
 import { COMMON_CONFIG } from '../../test-config';
 import { BehaviorSubject } from 'rxjs';
-import { skip, take, switchMap } from 'rxjs/operators';
+import { skip, switchMap, take } from 'rxjs/operators';
 import 'firebase/database';
 import { rando } from '../../firestore/utils.spec';
 
 describe('snapshotChanges', () => {
   let app: FirebaseApp;
   let db: AngularFireDatabase;
-  let createRef: (path: string) => database.Reference;
+  let createRef: (path: string) => firebase.database.Reference;
   let batch = {};
-  const items = [{ name: 'zero' }, { name: 'one' }, { name: 'two' }].map((item, i) => ( { key: i.toString(), ...item } ));
-  Object.keys(items).forEach(function (key, i) {
-    const itemValue = items[key];
-    batch[i] = itemValue;
+  const items = [{ name: 'zero' }, { name: 'one' }, { name: 'two' }].map((item, i) => ({ key: i.toString(), ...item }));
+  Object.keys(items).forEach((key, i) => {
+    batch[i] = items[key];
   });
   // make batch immutable to preserve integrity
   batch = Object.freeze(batch);
@@ -31,11 +30,10 @@ describe('snapshotChanges', () => {
         { provide: URL, useValue: 'http://localhost:9000' }
       ]
     });
-    inject([FirebaseApp, AngularFireDatabase], (app_: FirebaseApp, _db: AngularFireDatabase) => {
-      app = app_;
-      db = _db;
-      createRef = (path: string) => db.database.ref(path);;
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    db = TestBed.inject(AngularFireDatabase);
+    createRef = (path: string) => db.database.ref(path);
   });
 
   afterEach(() => {
@@ -55,7 +53,7 @@ describe('snapshotChanges', () => {
   it('should listen to all events by default', (done) => {
     const { snapChanges, ref } = prepareSnapshotChanges();
     snapChanges.pipe(take(1)).subscribe(actions => {
-      const data = actions.map(a => a.payload!.val());
+      const data = actions.map(a => a.payload.val());
       expect(data).toEqual(items);
     }).add(done);
     ref.set(batch);
@@ -63,9 +61,10 @@ describe('snapshotChanges', () => {
 
   it('should handle multiple subscriptions (hot)', (done) => {
     const { snapChanges, ref } = prepareSnapshotChanges();
-    const sub = snapChanges.subscribe(() => {}).add(done);
+    const sub = snapChanges.subscribe(() => {
+    }).add(done);
     snapChanges.pipe(take(1)).subscribe(actions => {
-      const data = actions.map(a => a.payload!.val());
+      const data = actions.map(a => a.payload.val());
       expect(data).toEqual(items);
     }).add(sub);
     ref.set(batch);
@@ -73,19 +72,20 @@ describe('snapshotChanges', () => {
 
   it('should handle multiple subscriptions (warm)', done => {
     const { snapChanges, ref } = prepareSnapshotChanges();
-    snapChanges.pipe(take(1)).subscribe(() => {}).add(() => {
+    snapChanges.pipe(take(1)).subscribe(() => {
+    }).add(() => {
       snapChanges.pipe(take(1)).subscribe(actions => {
-        const data = actions.map(a => a.payload!.val());
+        const data = actions.map(a => a.payload.val());
         expect(data).toEqual(items);
       }).add(done);
     });
     ref.set(batch);
   });
 
- it('should listen to only child_added events', (done) => {
+  it('should listen to only child_added events', (done) => {
     const { snapChanges, ref } = prepareSnapshotChanges({ events: ['child_added'], skipnumber: 0 });
     snapChanges.pipe(take(1)).subscribe(actions => {
-      const data = actions.map(a => a.payload!.val());
+      const data = actions.map(a => a.payload.val());
       expect(data).toEqual(items);
     }).add(done);
     ref.set(batch);
@@ -118,24 +118,22 @@ describe('snapshotChanges', () => {
   });
 
   it('should handle dynamic queries that return empty sets', done => {
-    const ITEMS = 10;
     let count = 0;
-    let firstIndex = 0;
-    let namefilter$ = new BehaviorSubject<number|null>(null);
+    const namefilter$ = new BehaviorSubject<number | null>(null);
     const aref = createRef(rando());
     aref.set(batch);
     namefilter$.pipe(switchMap(name => {
-      const filteredRef = name ? aref.child('name').equalTo(name) : aref
+      const filteredRef = name ? aref.child('name').equalTo(name) : aref;
       return snapshotChanges(filteredRef);
-    }),take(2)).subscribe(data => {
+    }), take(2)).subscribe(data => {
       count = count + 1;
       // the first time should all be 'added'
-      if(count === 1) {
+      if (count === 1) {
         expect(Object.keys(data).length).toEqual(3);
         namefilter$.next(-1);
       }
       // on the second round, we should have filtered out everything
-      if(count === 2) {
+      if (count === 2) {
         expect(Object.keys(data).length).toEqual(0);
       }
     }).add(done);
