@@ -1,9 +1,9 @@
-import { Observable, from } from 'rxjs';
+import { from, Observable } from 'rxjs';
 import { fromCollectionRef } from '../observable/fromRef';
-import { map, filter, scan, observeOn } from 'rxjs/operators';
-import { firestore } from 'firebase/app';
+import { filter, map, observeOn, scan } from 'rxjs/operators';
+import firebase from 'firebase/app';
 
-import { DocumentChangeType, Query, DocumentData, DocumentChangeAction } from '../interfaces';
+import { DocumentChangeAction, DocumentChangeType, DocumentData, Query } from '../interfaces';
 import { validateEventsArray } from '../collection/collection';
 import { docChanges, sortedChanges } from '../collection/changes';
 import { AngularFirestore } from '../firestore';
@@ -26,12 +26,10 @@ import { AngularFirestore } from '../firestore';
  * // Subscribe to changes as snapshots. This provides you data updates as well as delta updates.
  * fakeStock.valueChanges().subscribe(value => console.log(value));
  */
-export class AngularFirestoreCollectionGroup<T=DocumentData> {
+export class AngularFirestoreCollectionGroup<T= DocumentData> {
   /**
    * The constructor takes in a CollectionGroupQuery to provide wrapper methods
    * for data operations and data streaming.
-   * @param query
-   * @param afs
    */
   constructor(
     private readonly query: Query,
@@ -41,26 +39,24 @@ export class AngularFirestoreCollectionGroup<T=DocumentData> {
    * Listen to the latest change in the stream. This method returns changes
    * as they occur and they are not sorted by query order. This allows you to construct
    * your own data structure.
-   * @param events
    */
   stateChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
-    if(!events || events.length === 0) {
+    if (!events || events.length === 0) {
       return docChanges<T>(this.query, this.afs.schedulers.outsideAngular).pipe(
         this.afs.keepUnstableUntilFirst
       );
     }
     return docChanges<T>(this.query, this.afs.schedulers.outsideAngular)
       .pipe(
-        this.afs.keepUnstableUntilFirst,
         map(actions => actions.filter(change => events.indexOf(change.type) > -1)),
-        filter(changes =>  changes.length > 0)
+        filter(changes =>  changes.length > 0),
+        this.afs.keepUnstableUntilFirst
       );
   }
 
   /**
    * Create a stream of changes as they occur it time. This method is similar to stateChanges()
    * but it collects each event in an array over time.
-   * @param events
    */
   auditTrail(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     return this.stateChanges(events).pipe(scan((current, action) => [...current, ...action], []));
@@ -69,7 +65,6 @@ export class AngularFirestoreCollectionGroup<T=DocumentData> {
   /**
    * Create a stream of synchronized changes. This method keeps the local array in sorted
    * query order.
-   * @param events
    */
   snapshotChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     const validatedEvents = validateEventsArray(events);
@@ -86,16 +81,15 @@ export class AngularFirestoreCollectionGroup<T=DocumentData> {
     const fromCollectionRefScheduled$ = fromCollectionRef<T>(this.query, this.afs.schedulers.outsideAngular);
     return fromCollectionRefScheduled$
       .pipe(
-        this.afs.keepUnstableUntilFirst,
-        map(actions => actions.payload.docs.map(a => a.data()))
+        map(actions => actions.payload.docs.map(a => a.data())),
+        this.afs.keepUnstableUntilFirst
       );
   }
 
   /**
    * Retrieve the results of the query once.
-   * @param options
    */
-  get(options?: firestore.GetOptions) {
+  get(options?: firebase.firestore.GetOptions) {
     return from(this.query.get(options)).pipe(
       observeOn(this.afs.schedulers.insideAngular)
     );
