@@ -1,15 +1,15 @@
-import { Observable, from } from 'rxjs';
+import { from, Observable } from 'rxjs';
 import { fromCollectionRef } from '../observable/fromRef';
-import { map, filter, scan, observeOn } from 'rxjs/operators';
-import { firestore } from 'firebase/app';
+import { filter, map, observeOn, scan } from 'rxjs/operators';
+import firebase from 'firebase/app';
 
-import { DocumentChangeType, CollectionReference, Query, DocumentReference, DocumentData, DocumentChangeAction } from '../interfaces';
+import { CollectionReference, DocumentChangeAction, DocumentChangeType, DocumentData, DocumentReference, Query } from '../interfaces';
 import { docChanges, sortedChanges } from './changes';
 import { AngularFirestoreDocument } from '../document/document';
 import { AngularFirestore } from '../firestore';
 
 export function validateEventsArray(events?: DocumentChangeType[]) {
-  if(!events || events!.length === 0) {
+  if (!events || events.length === 0) {
     events = ['added', 'removed', 'modified'];
   }
   return events;
@@ -38,7 +38,7 @@ export function validateEventsArray(events?: DocumentChangeType[]) {
  * // Subscribe to changes as snapshots. This provides you data updates as well as delta updates.
  * fakeStock.valueChanges().subscribe(value => console.log(value));
  */
-export class AngularFirestoreCollection<T=DocumentData> {
+export class AngularFirestoreCollection<T= DocumentData> {
   /**
    * The constructor takes in a CollectionReference and Query to provide wrapper methods
    * for data operations and data streaming.
@@ -47,7 +47,6 @@ export class AngularFirestoreCollection<T=DocumentData> {
    * when you update data it is not updating data to the window of your query unless
    * the data fits the criteria of the query. See the AssociatedRefence type for details
    * on this implication.
-   * @param ref
    */
   constructor(
     public readonly ref: CollectionReference,
@@ -58,25 +57,23 @@ export class AngularFirestoreCollection<T=DocumentData> {
    * Listen to the latest change in the stream. This method returns changes
    * as they occur and they are not sorted by query order. This allows you to construct
    * your own data structure.
-   * @param events
    */
   stateChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
-    if(!events || events.length === 0) {
+    if (!events || events.length === 0) {
       return docChanges<T>(this.query, this.afs.schedulers.outsideAngular).pipe(
         this.afs.keepUnstableUntilFirst
       );
     }
     return docChanges<T>(this.query, this.afs.schedulers.outsideAngular).pipe(
-      this.afs.keepUnstableUntilFirst,
       map(actions => actions.filter(change => events.indexOf(change.type) > -1)),
-      filter(changes =>  changes.length > 0)
+      filter(changes =>  changes.length > 0),
+      this.afs.keepUnstableUntilFirst
     );
   }
 
   /**
    * Create a stream of changes as they occur it time. This method is similar to stateChanges()
    * but it collects each event in an array over time.
-   * @param events
    */
   auditTrail(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     return this.stateChanges(events).pipe(scan((current, action) => [...current, ...action], []));
@@ -85,7 +82,6 @@ export class AngularFirestoreCollection<T=DocumentData> {
   /**
    * Create a stream of synchronized changes. This method keeps the local array in sorted
    * query order.
-   * @param events
    */
   snapshotChanges(events?: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
     const validatedEvents = validateEventsArray(events);
@@ -100,33 +96,32 @@ export class AngularFirestoreCollection<T=DocumentData> {
    *
    * If the `idField` option is provided, document IDs are included and mapped to the
    * provided `idField` property name.
-   * @param options
    */
-  valueChanges(): Observable<T[]>
-  valueChanges({}): Observable<T[]>
-  valueChanges<K extends string>(options: {idField: K}): Observable<(T & { [T in K]: string })[]>
+  valueChanges(): Observable<T[]>;
+  // tslint:disable-next-line:unified-signatures
+  valueChanges({}): Observable<T[]>;
+  valueChanges<K extends string>(options: {idField: K}): Observable<(T & { [T in K]: string })[]>;
   valueChanges<K extends string>(options: {idField?: K} = {}): Observable<T[]> {
     return fromCollectionRef<T>(this.query, this.afs.schedulers.outsideAngular)
       .pipe(
-        this.afs.keepUnstableUntilFirst,
         map(actions => actions.payload.docs.map(a => {
           if (options.idField) {
             return {
-              ...a.data() as Object,
+              ...a.data() as {},
               ...{ [options.idField]: a.id }
             } as T & { [T in K]: string };
           } else {
-            return a.data()
+            return a.data();
           }
-        }))
+        })),
+        this.afs.keepUnstableUntilFirst
       );
   }
 
   /**
    * Retrieve the results of the query once.
-   * @param options
    */
-  get(options?: firestore.GetOptions) {
+  get(options?: firebase.firestore.GetOptions) {
     return from(this.query.get(options)).pipe(
       observeOn(this.afs.schedulers.insideAngular),
     );
@@ -145,7 +140,6 @@ export class AngularFirestoreCollection<T=DocumentData> {
 
   /**
    * Create a reference to a single document in a collection.
-   * @param path
    */
   doc<T>(path?: string): AngularFirestoreDocument<T> {
     return new AngularFirestoreDocument<T>(this.ref.doc(path), this.afs);
