@@ -1,29 +1,38 @@
-import { Injectable, Inject, Optional, NgZone, InjectionToken } from '@angular/core';
-import { of, from, Observable } from 'rxjs';
-import { map, switchMap, shareReplay, tap, observeOn } from 'rxjs/operators';
-import { FirebaseOptions, FirebaseAppConfig, FIREBASE_APP_NAME, ɵlazySDKProxy, ɵPromiseProxy, ɵAngularFireSchedulers } from '@angular/fire';
-import { FIREBASE_OPTIONS, ɵfirebaseAppFactory } from '@angular/fire';
-import { functions } from 'firebase/app';
+import { Inject, Injectable, InjectionToken, NgZone, Optional } from '@angular/core';
+import { from, Observable, of } from 'rxjs';
+import { map, observeOn, shareReplay, switchMap, tap } from 'rxjs/operators';
+import {
+  FIREBASE_APP_NAME,
+  FIREBASE_OPTIONS,
+  FirebaseAppConfig,
+  FirebaseOptions,
+  ɵAngularFireSchedulers,
+  ɵfirebaseAppFactory,
+  ɵlazySDKProxy,
+  ɵPromiseProxy
+} from '@angular/fire';
+import firebase from 'firebase/app';
 
 export const ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
 export const REGION = new InjectionToken<string>('angularfire2.functions.region');
 
 // override httpsCallable for compatibility with 5.x
-export interface AngularFireFunctions extends Omit<ɵPromiseProxy<functions.Functions>, 'httpsCallable'> { };
+export interface AngularFireFunctions extends Omit<ɵPromiseProxy<firebase.functions.Functions>, 'httpsCallable'> {
+}
 
 @Injectable({
   providedIn: 'any'
 })
 export class AngularFireFunctions {
 
-  public readonly httpsCallable: <T=any, R=any>(name: string) => (data: T) => Observable<R>
+  public readonly httpsCallable: <T = any, R = any>(name: string) => (data: T) => Observable<R>;
 
   constructor(
-    @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
-    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig:string|FirebaseAppConfig|null|undefined,
+    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
     zone: NgZone,
-    @Optional() @Inject(REGION) region:string|null,
-    @Optional() @Inject(ORIGIN) origin:string|null
+    @Optional() @Inject(REGION) region: string | null,
+    @Optional() @Inject(ORIGIN) origin: string | null
   ) {
     const schedulers = new ɵAngularFireSchedulers(zone);
 
@@ -33,17 +42,19 @@ export class AngularFireFunctions {
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
       map(app => app.functions(region || undefined)),
       tap(functions => {
-        if (origin) { functions.useFunctionsEmulator(origin) }
+        if (origin) {
+          functions.useFunctionsEmulator(origin);
+        }
       }),
-      shareReplay({ bufferSize: 1, refCount: false }),
+      shareReplay({ bufferSize: 1, refCount: false })
     );
 
-    this.httpsCallable = <T=any, R=any>(name: string) =>
+    this.httpsCallable = <T = any, R = any>(name: string) =>
       (data: T) => from(functions).pipe(
-        observeOn(schedulers.outsideAngular),
+        observeOn(schedulers.insideAngular),
         switchMap(functions => functions.httpsCallable(name)(data)),
         map(r => r.data as R)
-      )
+      );
 
     return ɵlazySDKProxy(this, functions, zone);
 
