@@ -1,12 +1,21 @@
-import { Injectable, Inject, Optional, NgZone, PLATFORM_ID } from '@angular/core';
-import { messaging } from 'firebase/app';
-import { Observable, empty, of, throwError } from 'rxjs';
-import { mergeMap, catchError, map, switchMap, concat, observeOn, defaultIfEmpty } from 'rxjs/operators';
-import { FirebaseOptions, FirebaseAppConfig, ɵAngularFireSchedulers, FIREBASE_APP_NAME, FIREBASE_OPTIONS, ɵlazySDKProxy, ɵPromiseProxy } from '@angular/fire';
-import { ɵfirebaseAppFactory } from '@angular/fire';
+import { Inject, Injectable, NgZone, Optional, PLATFORM_ID } from '@angular/core';
+import firebase from 'firebase/app';
+import { concat, EMPTY, Observable, of, throwError, fromEvent } from 'rxjs';
+import { catchError, defaultIfEmpty, map, mergeMap, observeOn, switchMap, switchMapTo, shareReplay, filter, subscribeOn } from 'rxjs/operators';
+import {
+  FIREBASE_APP_NAME,
+  FIREBASE_OPTIONS,
+  FirebaseAppConfig,
+  FirebaseOptions,
+  ɵAngularFireSchedulers,
+  ɵfirebaseAppFactory,
+  ɵlazySDKProxy,
+  ɵPromiseProxy
+} from '@angular/fire';
 import { isPlatformServer } from '@angular/common';
 
-export interface AngularFireMessaging extends Omit<ɵPromiseProxy<messaging.Messaging>, 'deleteToken'|'getToken'|'requestPermission'> {};
+export interface AngularFireMessaging extends Omit<ɵPromiseProxy<firebase.messaging.Messaging>, 'deleteToken' | 'getToken' | 'requestPermission'> {
+}
 
 @Injectable({
   providedIn: 'any'
@@ -14,62 +23,66 @@ export interface AngularFireMessaging extends Omit<ɵPromiseProxy<messaging.Mess
 export class AngularFireMessaging {
 
   public readonly requestPermission: Observable<void>;
-  public readonly getToken: Observable<string|null>;
-  public readonly tokenChanges: Observable<string|null>;
+  public readonly getToken: Observable<string | null>;
+  public readonly tokenChanges: Observable<string | null>;
   public readonly messages: Observable<{}>;
-  public readonly requestToken: Observable<string|null>;
+  public readonly requestToken: Observable<string | null>;
   public readonly deleteToken: (token: string) => Observable<boolean>;
 
   constructor(
-    @Inject(FIREBASE_OPTIONS) options:FirebaseOptions,
-    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig:string|FirebaseAppConfig|null|undefined,
+    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
+    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
+    // tslint:disable-next-line:ban-types
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone
   ) {
     const schedulers = new ɵAngularFireSchedulers(zone);
 
     const messaging = of(undefined).pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(() => isPlatformServer(platformId) ? empty() : import('firebase/messaging')),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(() => isPlatformServer(platformId) ? EMPTY : import('firebase/messaging')),
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
-      map(app => app.messaging())
+      map(app => app.messaging()),
+      shareReplay({ bufferSize: 1, refCount: false })
     );
 
-    if (!isPlatformServer(platformId)) {
-
-      this.requestPermission = messaging.pipe(
-        observeOn(schedulers.outsideAngular),
-        switchMap(messaging => messaging.requestPermission()),
-      );
-    
-    } else {
-    
-      this.requestPermission = throwError('Not available on server platform.');
-    
-    }
+    this.requestPermission = messaging.pipe(
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      // tslint:disable-next-line
+      switchMap(messaging => firebase.messaging.isSupported() ? messaging.requestPermission() : throwError('Not supported.'))
+    );
 
     this.getToken = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => messaging.getToken()),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() && Notification.permission === 'granted' ? messaging.getToken() : EMPTY),
       defaultIfEmpty(null)
-    )
+    );
 
     const tokenChanges = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => new Observable(messaging.onTokenRefresh.bind(messaging)).pipe(
-        switchMap(() => messaging.getToken())
-      ))
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? new Observable<string>(emitter =>
+        messaging.onTokenRefresh(emitter.next, emitter.error, emitter.complete)
+      ) : EMPTY),
+      switchMapTo(this.getToken)
     );
 
     this.tokenChanges = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => messaging.getToken()),
-      concat(tokenChanges)
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? concat(this.getToken, tokenChanges) : EMPTY)
     );
 
+
     this.messages = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => new Observable(messaging.onMessage.bind(messaging)))
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? new Observable<string>(emitter =>
+        messaging.onMessage(next => emitter.next(next), err => emitter.error(err), () => emitter.complete())
+      ) : EMPTY),
     );
 
     this.requestToken = of(undefined).pipe(
@@ -79,7 +92,8 @@ export class AngularFireMessaging {
     );
 
     this.deleteToken = (token: string) => messaging.pipe(
-      observeOn(schedulers.outsideAngular),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
       switchMap(messaging => messaging.deleteToken(token)),
       defaultIfEmpty(false)
     );
