@@ -1,4 +1,4 @@
-import { DatabaseQuery, AngularFireList, ChildEvent } from '../interfaces';
+import { AngularFireList, ChildEvent, DatabaseQuery } from '../interfaces';
 import { snapshotChanges } from './snapshot-changes';
 import { stateChanges } from './state-changes';
 import { auditTrail } from './audit-trail';
@@ -7,14 +7,15 @@ import { createRemoveMethod } from './remove';
 import { AngularFireDatabase } from '../database';
 import { map } from 'rxjs/operators';
 
-export function createListReference<T=any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireList<T> {
+export function createListReference<T= any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireList<T> {
   const outsideAngularScheduler = afDatabase.schedulers.outsideAngular;
+  const refInZone = afDatabase.schedulers.ngZone.run(() => query.ref);
   return {
     query,
-    update: createDataOperationMethod<Partial<T>>(query.ref, 'update'),
-    set: createDataOperationMethod<T>(query.ref, 'set'),
-    push: (data: T) => query.ref.push(data),
-    remove: createRemoveMethod(query.ref),
+    update: createDataOperationMethod<Partial<T>>(refInZone, 'update'),
+    set: createDataOperationMethod<T>(refInZone, 'set'),
+    push: (data: T) => refInZone.push(data),
+    remove: createRemoveMethod(refInZone),
     snapshotChanges(events?: ChildEvent[]) {
       return snapshotChanges<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
@@ -27,9 +28,9 @@ export function createListReference<T=any>(query: DatabaseQuery, afDatabase: Ang
     valueChanges(events?: ChildEvent[]) {
       const snapshotChanges$ = snapshotChanges<T>(query, events, outsideAngularScheduler);
       return snapshotChanges$.pipe(
-        afDatabase.keepUnstableUntilFirst,
-        map(actions => actions.map(a => a.payload.val() as T))
+        map(actions => actions.map(a => a.payload.val() as T)),
+        afDatabase.keepUnstableUntilFirst
       );
     }
-  }
+  };
 }
