@@ -1,13 +1,12 @@
-
-import { DatabaseReference, FirebaseOperation, DatabaseSnapshot } from '../interfaces';
+import { DatabaseReference, DatabaseSnapshot, FirebaseOperation } from '../interfaces';
 import { checkOperationCases } from '../utils';
 
 export function createDataOperationMethod<T>(ref: DatabaseReference, operation: string) {
   return function dataOperation<T>(item: FirebaseOperation, value: T) {
     return checkOperationCases(item, {
-      stringCase: () => ref.child(<string>item)[operation](value),
-      firebaseCase: () => (<DatabaseReference>item)[operation](value),
-      snapshotCase: () => (<DatabaseSnapshot<T>>item).ref[operation](value)
+      stringCase: () => ref.child(item as string)[operation](value),
+      firebaseCase: () => (item as DatabaseReference)[operation](value),
+      snapshotCase: () => (item as DatabaseSnapshot<T>).ref[operation](value)
     });
-  }
+  };
 }
