@@ -1,4 +1,3 @@
-import { TestBed, inject } from '@angular/core/testing';
 import * as utils from './utils';
 
 describe('utils', () => {
@@ -12,11 +11,12 @@ describe('utils', () => {
       const nul = null;
       const obj = {};
       const fn = () => { };
-      let undef;
+      const undef = undefined;
       expect(utils.isString(str)).toBe(true);
       expect(utils.isString(notStr)).toBe(false);
       expect(utils.isString(bool)).toBe(false);
       expect(utils.isString(nul)).toBe(false);
+      expect(utils.isString(obj)).toBe(false);
       expect(utils.isString(fn)).toBe(false);
       expect(utils.isString(undef)).toBe(false);
     });
