@@ -2,12 +2,11 @@ import { fromCollectionRef } from '../observable/fromRef';
 import { Observable, SchedulerLike } from 'rxjs';
 import { map, scan } from 'rxjs/operators';
 
-import { Query, DocumentChangeType, DocumentChange, DocumentChangeAction } from '../interfaces';
+import { DocumentChange, DocumentChangeAction, DocumentChangeType, Query } from '../interfaces';
 
 /**
  * Return a stream of document changes on a query. These results are not in sort order but in
  * order of occurence.
- * @param query
  */
 export function docChanges<T>(query: Query, scheduler?: SchedulerLike): Observable<DocumentChangeAction<T>[]> {
   return fromCollectionRef(query, scheduler)
@@ -19,9 +18,11 @@ export function docChanges<T>(query: Query, scheduler?: SchedulerLike): Observab
 
 /**
  * Return a stream of document changes on a query. These results are in sort order.
- * @param query
  */
-export function sortedChanges<T>(query: Query, events: DocumentChangeType[], scheduler?: SchedulerLike): Observable<DocumentChangeAction<T>[]> {
+export function sortedChanges<T>(
+  query: Query,
+  events: DocumentChangeType[],
+  scheduler?: SchedulerLike): Observable<DocumentChangeAction<T>[]> {
   return fromCollectionRef(query, scheduler)
     .pipe(
       map(changes => changes.payload.docChanges()),
@@ -32,14 +33,11 @@ export function sortedChanges<T>(query: Query, events: DocumentChangeType[], sch
 /**
  * Combines the total result set from the current set of changes from an incoming set
  * of changes.
- * @param current
- * @param changes
- * @param events
  */
 export function combineChanges<T>(current: DocumentChange<T>[], changes: DocumentChange<T>[], events: DocumentChangeType[]) {
   changes.forEach(change => {
     // skip unwanted change types
-    if(events.indexOf(change.type) > -1) {
+    if (events.indexOf(change.type) > -1) {
       current = combineChange(current, change);
     }
   });
@@ -48,11 +46,9 @@ export function combineChanges<T>(current: DocumentChange<T>[], changes: Documen
 
 /**
  * Creates a new sorted array from a new change.
- * @param combined
- * @param change
  */
 export function combineChange<T>(combined: DocumentChange<T>[], change: DocumentChange<T>): DocumentChange<T>[] {
-  switch(change.type) {
+  switch (change.type) {
     case 'added':
       if (combined[change.newIndex] && combined[change.newIndex].doc.ref.isEqual(change.doc.ref)) {
         // Not sure why the duplicates are getting fired
@@ -64,7 +60,7 @@ export function combineChange<T>(combined: DocumentChange<T>[], change: Document
       if (combined[change.oldIndex] == null || combined[change.oldIndex].doc.ref.isEqual(change.doc.ref)) {
         // When an item changes position we first remove it
         // and then add it's new position
-        if(change.oldIndex !== change.newIndex) {
+        if (change.oldIndex !== change.newIndex) {
           combined.splice(change.oldIndex, 1);
           combined.splice(change.newIndex, 0, change);
         } else {
