@@ -1,20 +1,18 @@
-import { FirebaseApp, AngularFireModule } from '@angular/fire';
+import { AngularFireModule, FirebaseApp } from '@angular/fire';
 import { AngularFirestore, SETTINGS } from '../firestore';
 import { AngularFirestoreModule } from '../firestore.module';
-import { AngularFirestoreDocument } from '../document/document';
-import { Observable, Subscription } from 'rxjs';
+import { AngularFirestoreDocument } from './document';
 import { take } from 'rxjs/operators';
 
-import { TestBed, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { COMMON_CONFIG } from '../../test-config';
 
-import { Stock, randomName, FAKE_STOCK_DATA, rando } from '../utils.spec';
+import { FAKE_STOCK_DATA, rando, randomName, Stock } from '../utils.spec';
 import 'firebase/firestore';
 
 describe('AngularFirestoreDocument', () => {
   let app: FirebaseApp;
   let afs: AngularFirestore;
-  let sub: Subscription;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -26,10 +24,9 @@ describe('AngularFirestoreDocument', () => {
         { provide: SETTINGS, useValue: { host: 'localhost:8080', ssl: false } }
       ]
     });
-    inject([FirebaseApp, AngularFirestore], (_app: FirebaseApp, _afs: AngularFirestore) => {
-      app = _app;
-      afs = _afs;
-    })();
+
+    app = TestBed.inject(FirebaseApp);
+    afs = TestBed.inject(AngularFirestore);
   });
 
   afterEach(() => {
@@ -59,7 +56,7 @@ describe('AngularFirestoreDocument', () => {
     await stock.set(FAKE_STOCK_DATA);
     const obs$ = stock.valueChanges();
     obs$.pipe(take(1)).subscribe(async data => {
-      expect(JSON.stringify(data)).toBe(JSON.stringify(FAKE_STOCK_DATA));
+      expect(data).toEqual(FAKE_STOCK_DATA);
       stock.delete().then(done).catch(done.fail);
     });
   });
