@@ -5,4 +5,4 @@ export * from './collection-group/collection-group';
 export * from './document/document';
 export * from './collection/changes';
 export * from './observable/fromRef';
-export * from './interfaces'
+export * from './interfaces';
