@@ -1,14 +1,9 @@
-import {
-  BuilderContext,
-  BuilderOutput,
-  createBuilder
-} from "@angular-devkit/architect";
-import { NodeJsSyncHost } from "@angular-devkit/core/node";
-import deploy from "./actions";
-import { experimental, normalize, json } from "@angular-devkit/core";
-import { DeployBuilderSchema } from '../interfaces';
-import * as path from "path";
-import { getFirebaseProjectName } from "../utils";
+import { BuilderContext, BuilderOutput, createBuilder } from '@angular-devkit/architect';
+import { NodeJsSyncHost } from '@angular-devkit/core/node';
+import deploy from './actions';
+import { experimental, json, normalize } from '@angular-devkit/core';
+import { BuildTarget, DeployBuilderSchema } from '../interfaces';
+import { getFirebaseProjectName } from '../utils';
 
 type DeployBuilderOptions = DeployBuilderSchema & json.JsonObject;
 
@@ -23,32 +18,50 @@ export default createBuilder<any>(
       new NodeJsSyncHost()
     );
     await workspace
-      .loadWorkspaceFromHost(normalize("angular.json"))
+      .loadWorkspaceFromHost(normalize('angular.json'))
       .toPromise();
 
     if (!context.target) {
-      throw new Error("Cannot deploy the application without a target");
+      throw new Error('Cannot deploy the application without a target');
     }
 
-    const project = workspace.getProject(context.target.project);
+    const projectTargets = workspace.getProjectTargets(context.target.project);
 
     const firebaseProject = getFirebaseProjectName(
       context.workspaceRoot,
       context.target.project
     );
 
+    if (!firebaseProject) {
+      throw new Error('Cannot find firebase project for your app in .firebaserc');
+    }
+
     const buildTarget = options.buildTarget || `${context.target.project}:build:production`;
+
+    const targets: BuildTarget[] = [{
+      name: buildTarget
+    }];
+    if (options.ssr) {
+      targets.push({
+        name: options.universalBuildTarget || `${context.target.project}:server:production`,
+        options: {
+          bundleDependencies: 'all'
+        }
+      });
+    }
 
     try {
       await deploy(
-        require("firebase-tools"),
+        require('firebase-tools'),
         context,
-        path.join(context.workspaceRoot, project.root),
-        buildTarget,
-        firebaseProject
+        projectTargets,
+        targets,
+        firebaseProject,
+        !!options.ssr,
+        !!options.preview
       );
     } catch (e) {
-      console.error("Error when trying to deploy: ");
+      console.error('Error when trying to deploy: ');
       console.error(e.message);
       return { success: false };
     }
