@@ -3,7 +3,7 @@ export interface Project {
   projectNumber: string;
   displayName: string;
   name: string;
-  resources: { [key:string]: string }
+  resources: { [key: string]: string };
 }
 
 export interface FirebaseDeployConfig {
@@ -12,31 +12,41 @@ export interface FirebaseDeployConfig {
 }
 
 export interface FirebaseTools {
-  login(): Promise<void>;
-
   projects: {
     list(): Promise<Project[]>;
-  }
+  };
+
+  logger: {
+    add(...args: any[]): any
+  };
+
+  login(): Promise<void>;
 
   deploy(config: FirebaseDeployConfig): Promise<any>;
+
+  serve(options: any): Promise<any>;
 
   use(options: any, lol: any): Promise<any>;
 }
 
 export interface FirebaseHostingRewrite {
   source: string;
-  destination: string;
+  destination?: string;
+  function?: string;
 }
 
 export interface FirebaseHostingConfig {
-  public: string;
+  public?: string;
   ignore: string[];
   target: string;
   rewrites: FirebaseHostingRewrite[];
 }
 
+export interface FirebaseFunctionsConfig { [key: string]: any; }
+
 export interface FirebaseJSON {
   hosting?: FirebaseHostingConfig[] | FirebaseHostingConfig;
+  functions?: FirebaseFunctionsConfig;
 }
 
 export interface FirebaseRcTarget {
@@ -49,4 +59,18 @@ export interface FirebaseRc {
 
 export interface DeployBuilderSchema {
   buildTarget?: string;
+  preview?: boolean;
+  universalBuildTarget?: string;
+  ssr?: boolean;
+}
+
+export interface BuildTarget {
+  name: string;
+  options?: {[name: string]: any};
+}
+
+export interface FSHost {
+  moveSync(src: string, dest: string): void;
+  writeFileSync(src: string, data: string): void;
+  renameSync(src: string, dest: string): void;
 }
