@@ -1,23 +1,23 @@
-import { SettableMetadata, UploadMetadata, Reference, StringFormat } from './interfaces';
-import { createUploadTask, AngularFireUploadTask } from './task';
-import { Observable, from } from 'rxjs';
+import { ListResult, Reference, SettableMetadata, StringFormat, UploadMetadata } from './interfaces';
+import { AngularFireUploadTask, createUploadTask } from './task';
+import { from, Observable, of } from 'rxjs';
 import { ɵAngularFireSchedulers } from '@angular/fire';
+import { observeOn, switchMap } from 'rxjs/operators';
 
 export interface AngularFireStorageReference {
   getDownloadURL(): Observable<any>;
   getMetadata(): Observable<any>;
   delete(): Observable<any>;
   child(path: string): any;
-  updateMetatdata(meta: SettableMetadata): Observable<any>;
   updateMetadata(meta: SettableMetadata): Observable<any>;
   put(data: any, metadata?: UploadMetadata | undefined): AngularFireUploadTask;
   putString(data: string, format?: string | undefined, metadata?: UploadMetadata | undefined): AngularFireUploadTask;
+  listAll(): Observable<ListResult>;
 }
 
 /**
  * Create an AngularFire wrapped Storage Reference. This object
  * creates observable methods from promise based methods.
- * @param ref
  */
 export function createStorageRef(
   ref: Reference,
@@ -25,15 +25,18 @@ export function createStorageRef(
   keepUnstableUntilFirst: <T>(obs$: Observable<T>) => Observable<T>
 ): AngularFireStorageReference {
   return {
-    getDownloadURL: () => from(ref.getDownloadURL(), schedulers.outsideAngular).pipe(
+    getDownloadURL: () => of(undefined).pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(() => ref.getDownloadURL()),
       keepUnstableUntilFirst
     ),
-    getMetadata: () => from(ref.getMetadata()).pipe(
+    getMetadata: () => of(undefined).pipe(
+      observeOn(schedulers.outsideAngular),
+      switchMap(() => ref.getMetadata()),
       keepUnstableUntilFirst
     ),
     delete: () => from(ref.delete()),
     child: (path: string) => createStorageRef(ref.child(path), schedulers, keepUnstableUntilFirst),
-    updateMetatdata: (meta: SettableMetadata) => from(ref.updateMetadata(meta)),
     updateMetadata: (meta: SettableMetadata) => from(ref.updateMetadata(meta)),
     put: (data: any, metadata?: UploadMetadata) => {
       const task = ref.put(data, metadata);
@@ -42,6 +45,7 @@ export function createStorageRef(
     putString: (data: string, format?: StringFormat, metadata?: UploadMetadata) => {
       const task = ref.putString(data, format, metadata);
       return createUploadTask(task);
-    }
+    },
+    listAll: () => from(ref.listAll())
   };
 }
