@@ -1,8 +1,8 @@
 import { spawn } from 'child_process';
-import { copy, writeFile, readFile } from 'fs-extra';
+import { copy, readFile, writeFile } from 'fs-extra';
 import { prettySize } from 'pretty-size';
 import { sync as gzipSync } from 'gzip-size';
-import { join, dirname } from 'path';
+import { dirname, join } from 'path';
 
 // TODO infer these from the package.json
 const MODULES = [
@@ -13,8 +13,8 @@ const MODULES = [
 const UMD_NAMES = MODULES.map(m => m === 'core' ? 'angular-fire' : `angular-fire-${m}`);
 const ENTRY_NAMES = MODULES.map(m => m === 'core' ? '@angular/fire' : `@angular/fire/${m}`);
 
-const src = (...args:string[]) => join(process.cwd(), 'src', ...args);
-const dest = (...args:string[]) => join(process.cwd(), 'dist', 'packages-dist', ...args);
+const src = (...args: string[]) => join(process.cwd(), 'src', ...args);
+const dest = (...args: string[]) => join(process.cwd(), 'dist', 'packages-dist', ...args);
 
 const rootPackage = import(join(process.cwd(), 'package.json'));
 
@@ -31,7 +31,7 @@ async function replacePackageCoreVersion() {
 async function replacePackageJsonVersions() {
   const path = dest('package.json');
   const root = await rootPackage;
-  var pkg = await import(path);
+  const pkg = await import(path);
   Object.keys(pkg.peerDependencies).forEach(peer => {
     pkg.peerDependencies[peer] = root.dependencies[peer];
   });
@@ -46,11 +46,14 @@ async function replaceSchematicVersions() {
   Object.keys(dependencies.default).forEach(name => {
     dependencies.default[name].version = root.dependencies[name] || root.devDependencies[name];
   });
+  Object.keys(dependencies.firebaseFunctions).forEach(name => {
+    dependencies.firebaseFunctions[name].version = root.dependencies[name] || root.devDependencies[name];
+  });
   return writeFile(path, JSON.stringify(dependencies, null, 2));
 }
 
 function spawnPromise(command: string, args: string[]) {
-  return new Promise(resolve => spawn(command, args, {stdio: 'inherit'}).on('close', resolve));
+  return new Promise(resolve => spawn(command, args, { stdio: 'inherit' }).on('close', resolve));
 }
 
 async function compileSchematics() {
@@ -90,30 +93,30 @@ function measureLibrary() {
 async function buildDocs() {
   // INVESTIGATE json to stdout rather than FS?
   await Promise.all(MODULES.map(module => spawnPromise('npx', ['typedoc', `./src/${module}`, '--json', `./dist/typedocs/${module}.json`])));
-  const entries = await Promise.all(MODULES.map(async (module, index) => {
+  const entries = await Promise.all(MODULES.map(async (module) => {
     const buffer = await readFile(`./dist/typedocs/${module}.json`);
     const typedoc = JSON.parse(buffer.toString());
     // TODO infer the entryPoint from the package.json
-    const entryPoint = typedoc.children.find((c:any) => c.name === "\"public_api\"");
+    const entryPoint = typedoc.children.find((c: any) => c.name === '"public_api"');
     const allChildren = [].concat(...typedoc.children.map(child =>
       // TODO chop out the working directory and filename
-      child.children ? child.children.map(c => ({...c, path: dirname(child.originalName.split(process.cwd())[1])})) : []
+      child.children ? child.children.map(c => ({ ...c, path: dirname(child.originalName.split(process.cwd())[1]) })) : []
     ));
     return entryPoint.children
       .filter(c => c.name[0] !== 'ɵ' && c.name[0] !== '_' /* private */)
-      .map(child => ({...allChildren.find(c => child.target === c.id)}))
-      .reduce((acc, child) => ({...acc, [encodeURIComponent(child.name)]: child}), {});
+      .map(child => ({ ...allChildren.find(c => child.target === c.id) }))
+      .reduce((acc, child) => ({ ...acc, [encodeURIComponent(child.name)]: child }), {});
   }));
   const root = await rootPackage;
   const pipes = ['MonoTypeOperatorFunction', 'OperatorFunction', 'AuthPipe', 'UnaryFunction'];
   const tocType = child => {
     const decorators: string[] = child.decorators && child.decorators.map(d => d.name) || [];
     if (decorators.includes('NgModule')) {
-      return 'NgModule'
+      return 'NgModule';
     } else if (child.kindString === 'Type alias') {
       return 'Type alias';
     } else if (child.kindString === 'Variable' && child.defaultValue && child.defaultValue.startsWith('new InjectionToken')) {
-      return 'InjectionToken'
+      return 'InjectionToken';
     } else if (child.type) {
       return pipes.includes(child.type.name) ? 'Pipe' : child.type.name;
     } else if (child.signatures && child.signatures[0] && child.signatures[0].type && pipes.includes(child.signatures[0].type.name)) {
@@ -121,12 +124,17 @@ async function buildDocs() {
     } else {
       return child.kindString;
     }
-  }
-  const table_of_contents = entries.reduce((acc, entry, index) =>
-    ({...acc, [MODULES[index]]: {name: ENTRY_NAMES[index], exports: Object.keys(entry).reduce((acc, key) => ({...acc, [key]: tocType(entry[key])}), {})}}),
+  };
+  const tableOfContents = entries.reduce((acc, entry, index) =>
+      ({
+        ...acc, [MODULES[index]]: {
+          name: ENTRY_NAMES[index],
+          exports: Object.keys(entry).reduce((acc, key) => ({ ...acc, [key]: tocType(entry[key]) }), {})
+        }
+      }),
     {}
   );
-  const afdoc = entries.reduce((acc, entry, index) => ({...acc, [MODULES[index]]: entry }), { table_of_contents });
+  const afdoc = entries.reduce((acc, entry, index) => ({ ...acc, [MODULES[index]]: entry }), { table_of_contents: tableOfContents });
   return writeFile(`./api-${root.version}.json`, JSON.stringify(afdoc, null, 2));
 }
 
@@ -141,6 +149,6 @@ Promise.all([
   console.log(`
 Package         Size    Gzipped
 ------------------------------------
-${stats.map((s, i) => [MODULES[i].padEnd(16), s.size.padEnd(8), s.gzip].join("")).join("\n")}`
+${stats.map((s, i) => [MODULES[i].padEnd(16), s.size.padEnd(8), s.gzip].join('')).join('\n')}`
   )
 );
